+++ tools/accuracy_checker/accuracy_checker/launcher/dlsdk_launcher_config.py
@@ -15,16 +15,22 @@ limitations under the License.
 """
 
 from pathlib import Path
+import platform
 from ..config import PathField, ConfigError, StringField, NumberField, ListField, DictField, BaseField, BoolField
 from .launcher import LauncherConfigValidator
-from .model_conversion import FrameworkParameters
+from .model_conversion import FrameworkParameters, convert_model
 from ..logging import warning
-from ..utils import get_path, contains_all
+from ..utils import get_path, contains_all, UnsupportedPackage, get_parameter_value_from_config
 
 try:
     from openvino.inference_engine import known_plugins
 except ImportError:
     known_plugins = []
+
+try:
+    from cpuinfo import get_cpu_info
+except ImportError as import_error:
+    get_cpu_info = UnsupportedPackage("cpuinfo", import_error.msg)
 
 HETERO_KEYWORD = 'HETERO:'
 MULTI_DEVICE_KEYWORD = 'MULTI:'
@@ -244,3 +250,91 @@ DLSDK_LAUNCHER_PARAMETERS = {
     '_prev_bitstream': PathField(optional=True, description="path to bitstream from previous run (FPGA only)"),
     '_model_is_blob': BoolField(optional=True, description='hint for auto model search')
 }
+
+
+def get_cpu_extension(cpu_extensions, selection_mode):
+    def get_cpu_extensions_list(file_format, base_name, selection_mode):
+        if not selection_mode:
+            default_cpu_extension = file_format.format(base_name)
+            extension_list = list(extensions_path.glob(default_cpu_extension))
+
+            if extension_list:
+                return extension_list
+
+            if isinstance(get_cpu_info, UnsupportedPackage):
+                get_cpu_info.raise_error("CPU extensions automatic search")
+
+            cpu_info_flags = get_cpu_info()['flags']
+            supported_flags = ['avx512', 'avx2', 'sse4_1', 'sse4_2']
+            cpu_info_flag_to_suffix = {
+                'avx512': 'avx512',
+                'avx2': 'avx2',
+                'sse4_1': 'sse4',
+                'sse4_2': 'sse4'
+            }
+            for flag in supported_flags:
+                selection_mode = cpu_info_flag_to_suffix[flag]
+                if flag in cpu_info_flags:
+                    break
+
+        extension_list = list(extensions_path.glob(file_format.format('{}_{}'.format(base_name, selection_mode))))
+        return extension_list
+
+    os_specific_formats = {
+        'Darwin': ('lib{}.dylib', 'lib{}.so'),
+        'Linux': ('lib{}.so',),
+        'Windows': ('{}.dll',),
+    }
+    cpu_extensions_name = cpu_extensions.parts[-1]
+    if cpu_extensions_name != 'AUTO':
+        return cpu_extensions
+    extensions_path = cpu_extensions.parent
+    system_name = platform.system()
+    file_formats = os_specific_formats.get(system_name)
+    if not file_formats:
+        raise ConfigError(
+            'Accuracy Checker can not automatically find cpu extensions library '
+            'for {} platform. Please, set cpu extension library manually.'.format(system_name)
+        )
+    extension_list = []
+    for supported_format in file_formats:
+        extension_list = get_cpu_extensions_list(supported_format, 'cpu_extension', selection_mode)
+        if extension_list:
+            break
+    if not extension_list:
+        raise ConfigError('suitable CPU extension lib not found in {}'.format(extensions_path))
+    return extension_list[0]
+
+
+def mo_convert_model(config, launcher_parameters, framework=None):
+    if framework is None:
+        framework = DLSDKLauncherConfigValidator.check_model_source(config)
+    config_model = config.get('{}_model'.format(framework.name), '')
+    config_weights = config.get('{}_weights'.format(framework.name), '')
+    config_meta = config.get('{}_meta'.format(framework.name), '')
+    mo_search_paths = []
+    model_optimizer = get_parameter_value_from_config(config, launcher_parameters, '_model_optimizer')
+    if model_optimizer:
+        mo_search_paths.append(model_optimizer)
+    model_optimizer_directory_env = os.environ.get('MO_DIR')
+    if model_optimizer_directory_env:
+        mo_search_paths.append(model_optimizer_directory_env)
+    model_name = (
+        Path(config_model).name.rsplit('.', 1)[0] or
+        Path(config_weights).name.rsplit('.', 1)[0] or
+        Path(config_meta).name.rsplit('.', 1)[0]
+    )
+    should_log_mo_cmd = get_parameter_value_from_config(config, launcher_parameters, 'should_log_cmd')
+
+    return convert_model(
+        model_name,
+        config_model, config_weights, config_meta, framework, mo_search_paths,
+        get_parameter_value_from_config(config, launcher_parameters, 'mo_params'),
+        get_parameter_value_from_config(config, launcher_parameters, 'mo_flags'),
+        get_parameter_value_from_config(config, launcher_parameters, '_tf_custom_op_config_dir'),
+        get_parameter_value_from_config(
+            config, launcher_parameters, '_tf_obj_detection_api_pipeline_config_path'
+        ),
+        get_parameter_value_from_config(config, launcher_parameters, '_transformations_config_dir'),
+        should_log_cmd=should_log_mo_cmd
+    )
